// frontend/app/api/generated/models/EmailVerificationChallenge.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Amra Merch Market API
 * Contract-first API for Amra Shop.
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface EmailVerificationChallenge
 */
export interface EmailVerificationChallenge {
    /**
     *
     * @type {string}
     * @memberof EmailVerificationChallenge
     */
    challengeId: string;
    /**
     *
     * @type {string}
     * @memberof EmailVerificationChallenge
     */
    email: string;
    /**
     *
     * @type {number}
     * @memberof EmailVerificationChallenge
     */
    expiresInSeconds: number;
    /**
     *
     * @type {string}
     * @memberof EmailVerificationChallenge
     */
    developmentCode?: string | null;
}

/**
 * Check if a given object implements the EmailVerificationChallenge interface.
 */
export function instanceOfEmailVerificationChallenge(value: object): value is EmailVerificationChallenge {
    if (!('challengeId' in value) || value['challengeId'] === undefined) return false;
    if (!('email' in value) || value['email'] === undefined) return false;
    if (!('expiresInSeconds' in value) || value['expiresInSeconds'] === undefined) return false;
    return true;
}

export function EmailVerificationChallengeFromJSON(json: any): EmailVerificationChallenge {
    return EmailVerificationChallengeFromJSONTyped(json, false);
}

export function EmailVerificationChallengeFromJSONTyped(json: any, ignoreDiscriminator: boolean): EmailVerificationChallenge {
    if (json == null) {
        return json;
    }
    return {

        'challengeId': json['challengeId'],
        'email': json['email'],
        'expiresInSeconds': json['expiresInSeconds'],
        'developmentCode': json['developmentCode'] == null ? undefined : json['developmentCode'],
    };
}

export function EmailVerificationChallengeToJSON(json: any): EmailVerificationChallenge {
    return EmailVerificationChallengeToJSONTyped(json, false);
}

export function EmailVerificationChallengeToJSONTyped(value?: EmailVerificationChallenge | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'challengeId': value['challengeId'],
        'email': value['email'],
        'expiresInSeconds': value['expiresInSeconds'],
        'developmentCode': value['developmentCode'],
    };
}